+++ CHANGELOG.md
@@ -4,8 +4,10 @@ Canonical reference for changes, improvements, and bugfixes for Boundary.
 
 ## Next
 
-## Bug Fixes
+### Bug Fixes
 
+* events: Fix panic when using the `hclog-text` event's format.
+  ([PR](https://github.com/hashicorp/boundary/pull/1456))
 * oidc managed groups: Allow colons in selector paths
   ([PR](https://github.com/hashicorp/boundary/pull/1453))